Convert vm details state to a controller
- reduced number of api calls through attribute consolidation
