+++ client/app/services/detail-reveal/detail-reveal.html
@@ -3,11 +3,11 @@
 		<div class="detail-reveal-arrow-container" ng-class="vm.hasMoreDetails ? '' : 'no-details' ">
 			<i class="fa fa-lg detail-reveal-arrow" ng-class="(vm.toggleDetails===true ? 'fa-angle-down' : 'fa-angle-right')"></i>
 		</div>
-		<span>{{::vm.title}}</span>
+		<span>{{vm.title}}</span>
 	</div>
 	<div class="col-xs-6 col-md-9">
-		<i ng-if="::vm.icon" class="{{::vm.icon}}"></i> 
-		<span>{{::vm.detail}}</span>
+		<i ng-if="vm.icon" class="{{vm.icon}}"></i> 
+		<span>{{vm.detail}}</span>
 	</div>
 </div>
 <div class="row item-extra-details-container" ng-if="vm.hasMoreDetails" uib-collapse="vm.toggleDetails===false">
