+++ client/app/services/vm-details/_vm-details.sass
@@ -1,3 +1,7 @@
+@import 'app_colors'
+@import 'patternfly/color-variables'
+
+
 .vm-details-wrapper
   padding: 0 20px
