+++ client/app/states/_states.sass
@@ -12,4 +12,3 @@
 @import 'dashboard/dashboard'
 @import 'orders/details/details'
 @import 'requests/details/details'
-@import 'vms/details/details'
